- Added min-max limits to numbers

+++ src/main/java/net/deadlydiamond98/koalalib/config/CFGProperties.java
@@ -13,6 +13,6 @@ import java.lang.annotation.Target;
 public @interface CFGProperties {
 
     boolean hasDesc() default true;
-    double min() default Double.MIN_VALUE;
-    double max() default Double.MAX_VALUE;
+    double min() default Integer.MIN_VALUE;
+    double max() default Integer.MAX_VALUE;
 }
